CustSupport_RAG/requirements.txt
# Core Dependencies

langchain>=0.2.0
langchain-core>=0.2.2
langchain-community>=0.2.0
langchain-google-genai>=1.0.6
langgraph>=0.2.34
langchain-text-splitters>=0.2.0

# Vector Store
faiss-cpu>=1.8.0


# Document Processing
pypdf==4.0.1
pymupdf==1.26.7  # For PyMuPDFLoader with table/image extraction
python-docx==1.1.0
pdf2image==1.17.0
Pillow==10.3.0
pytesseract==0.3.13
pandas==2.2.2
openpyxl==3.1.5

# Embeddings (for HuggingFace models like BAAI/bge-base-en-v1.5)
sentence-transformers>=2.5.1
transformers>=4.40.2
torch>=2.2.2

# API and Server
fastapi==0.110.1
uvicorn==0.30.1
pydantic>=2.9.0,<3.0.0
python-multipart==0.0.9

# UI
streamlit==1.33.0

# File Watching
watchdog==4.0.0

# Utilities
python-dotenv==1.0.1
numpy==1.26.4
tqdm==4.66.4

# Image Processing
opencv-python==4.9.0.80

# Table Extraction
tabula-py==2.9.0
camelot-py==0.11.0


